BibDocFile: error reporting changes

+++ modules/bibdocfile/lib/bibdocfile_webinterface.py
@@ -27,7 +27,8 @@ from invenio.config import \
      CFG_SITE_URL, \
      CFG_SITE_SECURE_URL, \
      CFG_WEBSUBMIT_STORAGEDIR, \
-     CFG_SITE_RECORD
+     CFG_SITE_RECORD, \
+     CFG_INSPIRE_SITE
 from invenio.bibdocfile_config import CFG_BIBDOCFILE_DOCUMENT_FILE_MANAGER_DOCTYPES, \
      CFG_BIBDOCFILE_DOCUMENT_FILE_MANAGER_MISC, \
      CFG_BIBDOCFILE_DOCUMENT_FILE_MANAGER_RESTRICTIONS, \
@@ -194,7 +195,7 @@ class WebInterfaceFilesPages(WebInterfaceDirectory):
                                 docfile = doc.get_file(docformat, version)
                             except InvenioBibDocFileError, msg:
                                 req.status = apache.HTTP_NOT_FOUND
-                                if req.headers_in.get('referer'):
+                                if not CFG_INSPIRE_SITE and req.headers_in.get('referer'):
                                     ## There must be a broken link somewhere.
                                     ## Maybe it's good to alert the admin
                                     register_exception(req=req, alert_admin=True)
@@ -381,17 +382,24 @@ class WebInterfaceManageDocFilesPages(WebInterfaceDirectory):
             working_dir = os.path.join(CFG_TMPSHAREDDIR,
                                        'websubmit_upload_interface_config_' + str(uid),
                                        argd['access'])
-            move_uploaded_files_to_storage(working_dir=working_dir,
-                                           recid=argd['recid'],
-                                           icon_sizes=['180>','700>'],
-                                           create_icon_doctypes=['*'],
-                                           force_file_revision=False)
-            # Clean temporary directory
-            shutil.rmtree(working_dir)
+            if not os.path.isdir(working_dir):
+                # We accessed the url without preliminary steps
+                # (we did not upload a file)
+                # Our working dir does not exist
+                # Display the file manager
+                argd['do'] = 0
+            else:
+                move_uploaded_files_to_storage(working_dir=working_dir,
+                                               recid=argd['recid'],
+                                               icon_sizes=['180>', '700>'],
+                                               create_icon_doctypes=['*'],
+                                               force_file_revision=False)
+                # Clean temporary directory
+                shutil.rmtree(working_dir)
 
-            # Confirm modifications
-            body += '<p style="color:#0f0">%s</p>' % \
-                    (_('Your modifications to record #%i have been submitted') % argd['recid'])
+                # Confirm modifications
+                body += '<p style="color:#0f0">%s</p>' % \
+                        (_('Your modifications to record #%i have been submitted') % argd['recid'])
         elif argd['cancel']:
             # Clean temporary directory
             working_dir = os.path.join(CFG_TMPSHAREDDIR,
